I think batch rendering is truly dynamic now.
Can change position and scale of the quad.

Each quad is represented as a Mesh. A mesh contains information about the translate, rotation and scale. Then it has a vector of VertexWrapper.
A VertexWrapper holds the vertex's position, and a Vertex struct. The Vertex struct holds the information that will be uploaded to the GPU